+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-12-11
+### Added
+- Added `"symfony/dependency-injection": "^5.0"` and `"symfony/config": "^5.0"` dependencies for Symfony 5 support;
+- Added `"php": "^7.2.5"` requirement.
+
+### Removed
+- Dropped support for Symfony `^2.3|^3.0|^4.0` due to incompatible interface changes.
+
 ## [1.1.0]
 ### Added
 - `AddTaggedCompilerPass` support for parameters with default values (associative array);